# tests/EventBooking.Web.Tests/Contracts/openapi-v1.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "EventBooking API",
    "description": "EventBooking staff API and anonymous Attendee booking links.",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:5001/"
    }
  ],
  "paths": {
...
    "/api/staff-access": {
      "get": {
        "tags": [
          "Staff Access"
        ],
        "summary": "List staff access profiles.",
        "description": "Reads every profile with its display name, staff number, read-only roles and scope.",
        "operationId": "listStaffAccess",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageOfStaffAccessResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          { }
        ],
        "x-eventbooking-list": true,
        "x-capability": "ManageStaffAccess",
        "x-mcp-tool": "list_staff_access",
        "x-agent-hints": {
          "readOnly": true,
          "destructive": false,
          "idempotent": true,
          "openWorld": false
        }
      }
    },
    "/api/staff-access/{staffUserId}/scope": {
      "put": {
        "tags": [
          "Staff Access"
        ],
        "summary": "Set a staff access scope.",
        "description": "Sets or clears one profile's appointment-type scope. Assigning a type that another Manager holds displaces them, and the response names who.",
        "operationId": "setStaffAccessScope",
        "parameters": [
          {
            "name": "staffUserId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JsonElement"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SetStaffScopeOutcome"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          { }
        ],
        "x-capability": "ManageStaffAccess",
        "x-mcp-tool": "set_staff_access_scope",
        "x-agent-hints": {
          "readOnly": false,
          "destructive": true,
          "idempotent": true,
          "openWorld": false
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiLink": {
        "required": [
          "href",
          "method",
          "operationId"
        ],
        "type": "object",
        "properties": {
          "href": {
            "type": "string"
          },
          "method": {
            "type": "string"
          },
          "operationId": {
            "type": "string"
          }
        }
      },
...
      "JsonElement": { },
...
      "PageOfStaffAccessResponse": {
        "required": [
          "items",
          "nextCursor"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StaffAccessResponse"
            }
          },
          "nextCursor": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "SetStaffScopeOutcome": {
        "required": [
          "targetStaffUserId",
          "appointmentTypeId",
          "displacedManagerDisplayName"
        ],
        "type": "object",
        "properties": {
          "targetStaffUserId": {
            "type": "string",
            "description": "The profile whose scope changed.",
            "format": "uuid"
          },
          "appointmentTypeId": {
            "type": [
              "null",
              "string"
            ],
            "description": "The scope now set, or null when cleared.",
            "format": "uuid"
          },
          "displacedManagerDisplayName": {
            "type": [
              "null",
              "string"
            ],
            "description": "The displaced holder's name, when any."
          }
        },
        "description": "The scope change plus the displaced Manager's display name when any."
      },
...
      "StaffAccessResponse": {
        "required": [
          "staffUserId",
          "staffId",
          "displayName",
          "roles",
          "appointmentTypeId",
          "version",
          "_links"
        ],
        "type": "object",
        "properties": {
          "staffUserId": {
            "type": "string",
            "format": "uuid"
          },
          "staffId": {
            "type": [
              "null",
              "string"
            ]
          },
          "displayName": {
            "type": [
              "null",
              "string"
            ]
          },
          "roles": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "appointmentTypeId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "version": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "_links": {
            "type": "object",
            "additionalProperties": {
              "$ref": "#/components/schemas/ApiLink"
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}